requirements.txt
playwright>=1.44.0
Pillow>=11.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
flask>=3.0.0
